requirements.txt
asgiref
Django
djangorestframework
Pygments
pytz
sqlparse
django-environ
psycopg2-binary